[tensor] Extract CpuComputeOps into its own header
CpuComputeOps was defined inside cpu_ops_table.cpp, so nothing outside
that translation unit could name it -- including an accelerator backend
that wants to inherit it. An accelerator typically overrides a handful
of ops and needs the other eighty to keep working; the base ComputeOps
defaults throw, so inheriting CpuComputeOps is what makes that possible.
Today ClComputeOps derives from ComputeOps and accepts the throw for
everything it does not implement, which is a fine choice for the GPU but
not the only one a backend should be able to make.

This is a pure move: the class declaration goes to cpu_ops_table.h and
cpu_ops_table.cpp keeps only get_cpu_ops(). No method body is edited, no
signature changes, and nothing new is compiled -- every one of the 338
moved lines is byte-identical to what it replaced. The header joins the
installed set (meson, debian, spec) because a downstream backend needs it
at build time.

**Self evaluation:**
1. Build test: [X]Passed [ ]Failed [ ]Skipped
2. Run test: [X]Passed [ ]Failed [ ]Skipped
