+++ debian/nntrainer-dev.install
@@ -28,6 +28,7 @@
 /usr/include/nntrainer/x86_compute_backend.h
 /usr/include/nntrainer/cpu_backend.h
 /usr/include/nntrainer/compute_ops.h
+/usr/include/nntrainer/cpu_ops_table.h
 /usr/include/nntrainer/context_data.h
 /usr/include/nntrainer/var_grad.h
 /usr/include/nntrainer/weight.h
