+++ nntrainer/tensor/cpu_backend/meson.build
@@ -1,6 +1,7 @@
 cpu_backend_headers = [
   'cpu_backend.h',
   'compute_ops.h',
+  'cpu_ops_table.h',
 ]
 
 nntrainer_sources += meson.current_source_dir() / 'compute_ops.cpp'
